Install golang from the package manager instead of external binaries
There is a golang cookbook but this also installs from external
binaries or from source, which seems pointless when all the supported
platforms already have a sufficiently recent version available. If
it's not in the main repositories then it can at least be found in
EPEL (which we already use) or Universe. The Ubuntu versions mentioned
in .kitchen.yml should be updated anyway.