Install PostgreSQL server in postgres recipe and connect locally
It's strange that we install a MySQL server in the mysql recipe but
don't install a PostgreSQL server in the postgres recipe.

This allows us to treat the server as local when creating the user and
database, simplifying the recipe.

The recipe no longer references the userhost attribute as this concept
does not exist in PostgreSQL.

Recent GitLab versions require the pg_trgm extension, which we now
install via the postgresql::contrib recipe. This really should be a
resource but we can force override node['postgresql']['database_name']
in the short term. We could install the extension to template1 instead
but this will not help existing users.