Leave GitLab to write its own secrets.yml file
As things stand, we are missing two additional keys from the
file. GitLab automatically adds these each time it starts but they
then get splatted by Chef on the next run, only to be reinstated with
different values later.

On top of this, the production_db_key_base attribute was not
documented so I have ended up with a value of "production" in my live
environment. I suspect others have too. This value should have been
randomly generated. Fortunately we were not using any features that
depend on this key.

Although it would be nice to preserve these keys in Chef for
migrations and restorations, this should be done securely using
encrypted data bags or similar. Until that is done, I believe it is
safer to just let GitLab generate its own keys. Perhaps this wasn't an
option before. This also avoids the list of keys falling behind.