Connect to Redis via a UNIX socket rather than TCP
We are already assuming that Redis is local. A UNIX socket is faster,
more secure, and avoids the IPv6 headache.

We now have to create the git user earlier so that the redisio
cookbook does not create it with different properties and because
usermod will fail if Redis is already running.